I found this type-stable but allocating version of the Gauss-Hermite `expected_loglikelihood` to be better for AD, compared to the original version and other type-stable implementations I tried, e.g. a hand-written `pairwise_sum(f, itr1, itr2)` (necessary because the multi-argument `mapreduce` does not use pairwise summation, see #77).
I also simplified the parameter conversions in `NegativeBinomialLikelihood(NBParamII(...))`.

This PR includes new tests that check the type stability of `expected_loglikelihood` with all integration methods and increases the number of tested likelihoods.

Closes #77