@simsurace 2022-06-22
Some benchmarks. This is the code to reproduce the example:

```julia
using BenchmarkTools
using Distributions
using GPLikelihoods
using Zygote

function testfun_small(θ)
    lik = NegativeBinomialLikelihood(NBParamII(θ[end]))
    q_f = Normal.(θ[1:3], θ[4:6])
    y = rand(0:10, 3)
    gh = GPLikelihoods.GaussHermiteExpectation(100)
    return expected_loglikelihood(gh, lik, q_f, y)
end
θ_small = rand(7)

function testfun_large(θ)
    lik = NegativeBinomialLikelihood(NBParamII(θ[end]))
    q_f = Normal.(θ[1:500], θ[501:1000])
    y = rand(0:10, 500)
    gh = GPLikelihoods.GaussHermiteExpectation(100)
    return expected_loglikelihood(gh, lik, q_f, y)
end
θ_large = rand(1001)

@btime testfun_small($θ_small)
@btime Zygote.gradient(testfun_small, $θ_small)

@btime testfun_large($θ_large)
@btime Zygote.gradient(testfun_large, $θ_large)
```
Here are the results for the small example, forward pass / reverse pass:
| Distributions         | GPLikelihoods#master     | [GPLikelihoods#ss/type_stability](https://github.com/simsurace/GPLikelihoods.jl/tree/ss/type_stability) | This PR |
|--------------|-----------|------------|--------------------|
| v0.25.62        | 120μs / 4698μs<br />(87: 35K / 61179: 2.3M)   | 119μs / 5569μs<br />(87: 35K / 13068: 1.2M)    | 122μs / 4345μs<br />(88: 37K / 48415: 1.8M) |
| [#1568](https://github.com/JuliaStats/Distributions.jl/pull/1568 )    | 120μs / 943μs<br />(91: 35K / 73409: 3.1M)     | 118μs / 249μs<br />(87: 38K / 823: 252K)      | 124μs / 179μs<br />(88: 37K / 396: 169K)  |

And here for the large example, forward pass / reverse pass:
| Distributions         | GPLikelihoods#master     | [GPLikelihoods#ss/type_stability](https://github.com/simsurace/GPLikelihoods.jl/tree/ss/type_stability) | This PR |
|--------------|-----------|------------|--------------------|
| v0.25.62        | 3.883ms / 1.203s<br />(91: 55K / 12182325: 507M)   | 3.965ms / 984ms<br />(87: 55K / 10150949: 387M)     | 4.422ms / 830ms<br />(89: 445K / 8004418: 289M) |
| [#1568](https://github.com/JuliaStats/Distributions.jl/pull/1568 )      | 4.372ms / 143ms<br />(90: 55K / 2125823: 184M)      | 4.422ms / 830ms<br />(87: 55K / 91450: 50M)      | 4.366ms / 8.880ms<br />(89: 445K / 4397: 19M)    |

Thus in the larger example the allocations become more significant for the forward pass. But it is still a clear win for the reverse pass.