@devmotion 2022-06-22
> Thus in the larger example the allocations become more significant for the forward pass. But it is still a clear win for the reverse pass.

Isn't that an argument for defining an `rrule`?