@simsurace 2022-06-22
I think the rrule that was clearly missing is the one from https://github.com/JuliaStats/Distributions.jl/pull/1568. That alone already moves the reverse pass into a more reasonable range. The question is why that is not sufficient. My working hypothesis was that it was due to the type-instability, so I pushed in that direction. I have now sunk a lot of time into this, but I think it has paid off. AD speed is now very well within the reasonable range. This is a complete game changer for fast training of GPs with count observations. The additional allocations in the forward pass don't matter that much for that application (I think in the data size where they would, you would anyway use sparse approximations).

I don't really have fundamental arguments against an even better/simpler solution, but I don't really much time to invest in this right now, given that I already spent that much. What function would you write the `rrule` for? The method for `expected_loglikelihood` that's being called here just contains a `Broadcast.broadcasted`. Would that need an `rrule`? It's unclear to be how one would write such a rule.

Unless a very trivial solution comes to mind, perhaps we can merge this and if the forward pass allocations become an issue later, we can always revisit this.