@devmotion 2022-06-22
> What function would you write the `rrule` for? The method for `expected_loglikelihood` that's being called here just contains a `Broadcast.broadcasted`. Would that need an `rrule`? It's unclear to be how one would write such a rule.

For `expected_loglikelihood`, similar to e.g. the rrule for `sum(f, x)` (https://github.com/JuliaDiff/ChainRules.jl/blob/7904019b89f9440925dc8a88ad77e483ebe81619/src/rulesets/Base/mapreduce.jl#L54-L107, without the special optimizations initially I assume). One could just collect the primal values and pullbacks of the individual `expected_loglikelihood` calls, as in https://github.com/JuliaDiff/ChainRules.jl/blob/7904019b89f9440925dc8a88ad77e483ebe81619/src/rulesets/Base/mapreduce.jl#L80-L81. I.e., one would also allocate - but only if one wants to differentiate. And in the pullback one collects and sums everything: https://github.com/JuliaDiff/ChainRules.jl/blob/7904019b89f9440925dc8a88ad77e483ebe81619/src/rulesets/Base/mapreduce.jl#L95-L103.