@simsurace 2022-07-03
Are there any blockers for this? It would be great to get this merged, as together with https://github.com/JuliaStats/Distributions.jl/pull/1568 (or its follow-up) this is a real game-changer for performance.