@simsurace 2022-07-07
Yes I agree it is a hack. I'll open an issue